pyproject.toml
[project]
name = "polar-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "charset-normalizer>=3.4.4",
    "polars-lts-cpu==1.33.1",
    "pymssql>=2.3.11",
    "pyodbc>=5.3.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
]
